fix(groups): `wm_class` with suffix follow default
The `wm_class` string of certain programs include the current version as
a suffix. This would break the default group assignment when using the
hash lookup.
The performance drop from iterating over the keys of
`wm_class_default_group` is negligible since the size is highly unlikely
to group significantly in size.